+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## next
+- add incremental parsing
+- bugfix: only create a new layer when extruding
+- allow input to be either a string or an array of strings
+- demo: shows how to use incremental parsing to render a file progressively
+
 ## 2.0
 - updated demo page
 - toggle rendering of travel